+++ frontend/index.html
@@ -3,7 +3,7 @@
 <head>
     <meta charset="UTF-8">
     <title>JS-Chat</title>
-    <script src="script.js"></script>
+    <script type="module" src="script.js"></script>
 </head>
 <body>
     <h1>JS-Chat</h1>
